Make "{{" OpenSpecial optional
You can now disable the special behaviour that makes typing "{{" convert
that second "{" into an indented line before the "}", by adding this to
your vimrc:
  let g:autoclose_openspecial = 0
This can be done globally, or in augroups in order to have it apply only
to specified file types/extensions.
Re-enabling the behaviour can be accomplished by leaving out the above
setting, or by setting it to 1.